+++ src/main/java/com/vinculacion/jpa/repository/TelefonoRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 /**
  * Created by Mauricio on 05/06/2017.
  */
-public interface TelefonoRepository extends JpaRepository<Telefono,Long>, Serializable {
-    Telefono findBytlfId(Long id);
-    List<Telefono> findByEstablecimiento_estId(Long id);
+public interface TelefonoRepository extends JpaRepository<Telefono,Integer>, Serializable {
+    Telefono findBytlfId(int id);
+    List<Telefono> findByEstablecimiento_estId(int id);
 }
